Revert "Added New structure to the website,"
