--- src/app/components/Main.module.css
@@ -1,10 +1,0 @@
-.main {
-    @apply bg-white px-5 md:px-20 lg:px-40 dark:bg-gray-900;
-  }
-
-  @font-face {
-    font-family: "Cedarville";
-    src: url("../../../public/CedarvilleCursive-Regular.ttf") format('truetype');
-    font-weight: normal;
-    font-style: normal;
-  }
