--- src/app/components/Sections/Hero.module.css
@@ -1,19 +1,0 @@
-.conv {
-    @apply text-center text-gray-500 space-y-0 dark:text-gray-400;
-}
-
-.name {
-    @apply text-5xl text-teal-600 space-y-0 font-medium md:text-6xl;
-}
-
-.summary {
-    @apply text-md leading-8 text-gray-800 md:text-xl lg:max-w-2xl max-w-lg mx-auto dark:text-gray-50;
-}
-
-.social_icons {
-    @apply justify-center flex text-5xl gap-x-16 py-5 text-gray-600 dark:text-gray-500;
-}
-
-.profilePicture { 
-    @apply relative rounded-full w-80 h-80 my-10 mx-auto overflow-hidden md:w-96 md:h-96;
-}
