--- src/app/components/Sections/Projects.module.css
@@ -1,8 +1,0 @@
-  .projects_containers {
-    @apply flex grid grid-cols-1 gap-4 dark:text-gray-200;
-  }
-
-  .project_title {
-    @apply m-4 text-center text-gray-700 dark:text-gray-100;
-  }
-
