--- src/app/components/Sections/Skills.module.css
@@ -1,11 +1,0 @@
-.skills_containers {
-    @apply flex grid grid-cols-1 gap-4 dark:text-gray-200 md:grid-cols-2 lg:grid-cols-3;
-  }
-
-.skills_title {
-    @apply text-teal-600 text-xl mb-4 font-medium;
-  }
-
-  .skills_subtitle {
-    @apply font-medium text-left my-4 dark:text-gray-100;
-  }
