+++ src/app/layout.js
@@ -4,6 +4,7 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 
+
 export const metadata = {
   title: "Mauricio Meza Portfolio",
   description: "Generated by create next app",
