+++ src/app/page.js
@@ -1,13 +1,319 @@
 'use client'
-import React from "react";
-import { DarkModeProvider } from "@/app/components/context/DarkModeContext";
-import Main from "./components/Main";
+import Head from "next/head";
+import Image from "next/image";
+
+import { useEffect, useState } from "react";
+
+import { BsFillMoonStarsFill, BsFillSunFill } from "react-icons/bs";
+import { AiFillGithub, AiFillLinkedin, AiFillInstagram } from "react-icons/ai";
+import { IoIosCheckmarkCircle } from "react-icons/io";
+
+import maudev from "/public/mau-meza.jpg";
+import web1 from "/public/web1.png";
+import web2 from "/public/mark-konig-workinprogress.jpg"
+
+import styles from "./page.module.css";
 
 export default function Home() {
-  return (
-    <DarkModeProvider>
-      <Main />
-    </DarkModeProvider>
-  );
-}
+  const [darkMode, setDarkMode] = useState(false);
+  const [icon, setIcon] = useState(<BsFillSunFill />);
+  
 
+  useEffect(() => {
+    setIcon(darkMode ? <BsFillSunFill color="white" /> : <BsFillMoonStarsFill />);
+  }, [darkMode]);
+
+  const toggleDarkMode = () => {
+    setDarkMode(!darkMode);
+  };
+
+  return (
+    <div className={darkMode ? "dark" : ""}>
+      <Head>
+        <title>Mauricio Meza Portfolio</title>
+        <meta name="Description" content="Generated by create next app"></meta>
+        <link rel='icon' href='./portfolio-32.png' />
+      </Head>
+
+      <main className={styles.main}>
+        <section className=" min-h-screen">
+          <nav className=" navigation">
+            <h1 className={styles.header}>
+              Mauricio Meza Dev.
+            </h1>
+            <ul className="flex items-center">
+              <ul className="flex items-center">
+                <a
+                  onClick={toggleDarkMode}
+                  className="cursor-pointer text-2xl"
+                >
+                  {icon}
+                </a>
+              </ul>
+              <li>
+                <a
+                  href="https://docs.google.com/document/d/1-dLvyX3TRGO-UIzUJhaP2SBk-xyW3D64/edit?usp=sharing&ouid=105066509412189301443&rtpof=true&sd=true" 
+                  className=" px-4 py-2 rounded-md ml-8 btn"
+                >
+                  Resume
+                </a>
+              </li>
+            </ul>
+          </nav>
+          <div className=" text-center p-5">
+
+            <div className=" ">
+              <p className={styles.conv}>I am</p>
+              <h1 className={styles.name}>
+                Mauricio Meza
+              </h1>
+            </div>
+            <p className={styles.summary}>
+              a skilled <span className="text-teal-500">Full-Stack developer</span> with a rich background in Mechanical Design and Manufacturing, boasting 7 years of engineering excellence, my journey in software development has been a fascinating transition. With experience as a <span className="text-teal-500"> Back-End developer</span> in my last job, I've honed my abilities in creating robust and efficient systems.
+            </p>
+            {/* <p className={styles.summary}>
+              From conceptualizing intricate designs to crafting elegant code, I thrive on the synergy of creativity and precision. I'm excited to share my experiences and creations with you as I navigate the realms of software development, combining my <span className="text-teal-500">engineering background</span> with a passion for building innovative solutions.
+            </p> */}
+          </div>
+          <div className={styles.social_icons}>
+            <a href="https://www.linkedin.com/in/mauricio-meza/?locale=en_US">
+              <AiFillLinkedin />
+            </a>
+            <a href="https://github.com/cs-mmeza">
+              <AiFillGithub />
+            </a>
+            <a href="https://www.instagram.com/maumeza17/">
+              <AiFillInstagram />
+            </a>
+          </div>
+          <div className={styles.profilePicture}>
+            <Image 
+              src={maudev} 
+              alt=""
+              fill
+              style={{objectFit:"cover"}} 
+            />
+          </div>
+          <div className="text-center p-5">
+          <p className={styles.summary}>
+              From conceptualizing intricate designs to crafting elegant code, I thrive on the synergy of creativity and precision. I'm excited to share my experiences and creations with you as I navigate the realms of software development, combining my <span className="text-teal-500">engineering background</span> with a passion for building innovative solutions.
+          </p>
+          </div>
+          
+        </section>
+        <section className="relative" id="skills">
+          <div className=" mt-10">
+          <p className="text-center text-gray-500">Check my</p>
+          <h2 className={styles.title}>
+            Tech Stack
+          </h2>
+          </div>
+          <div className="flex justify-center flex-col lg:flex gap-10">
+            <div className={styles.experience_details_container}>
+              <div className={styles.skills_containers}>
+                <div className="cards">
+                  <h3 className={styles.experience_title}>Front-End Development</h3>
+                  <div>
+                    <h4 className={styles.experience_subtitle}>Languages & Frameworks:</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>HTML</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>CSS</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>JavaScript</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>React</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Tailwindcss</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>DOM</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>TypeScript</h4>
+                    </article>
+                  </div>
+                </div>
+                <div className="cards">
+                  <h3 className={styles.experience_title}>Back-End Development</h3>
+                  <div>
+                    <h4 className={styles.experience_subtitle}>Languages & Frameworks:</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Express.js</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Python</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Django</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Java</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Spring</h4>
+                    </article>
+                    <h4 className={styles.experience_subtitle}>Version Control:</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Git</h4>
+                    </article>
+                    <h4 className={styles.experience_subtitle}>Back-End Development:</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>RESTfull APIs</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Microservices</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Docker</h4>
+                    </article>
+                  </div>
+                </div>
+                <div className="cards">
+                  <h3 className={styles.experience_title}>Data Processing & Analysis</h3>
+                  <div>
+                    <h4 className={styles.experience_subtitle}>Database Management</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>PostgreSQL</h4>
+                    </article>
+                    <article className="flex text-md text-left">
+                      <IoIosCheckmarkCircle className="flex mr-2" alt="experience icon" />
+                      <h4 className="">Relational Database</h4>
+                    </article>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>BigQuery</h4>
+                    </article>
+                    <h4 className={styles.experience_subtitle}>Data Processing</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>Data Cleaning</h4>
+                    </article>
+                    <article className="flex text-md text-left">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4 className="">Pattern Identification</h4>
+                    </article>
+                    <h4 className={styles.experience_subtitle}>Data Visualization</h4>
+                    <article className="flex">
+                      <IoIosCheckmarkCircle className="mr-2" alt="experience icon" />
+                      <h4>TableaU</h4>
+                    </article>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </section>
+        <section className="relative" id="portfolio">
+          <div className={styles.head}>
+            <p className="text-center text-gray-500">Explore my recent</p>
+            <h2 className={styles.title}>Projects</h2>
+          </div>
+          <div className={styles.experience_details_container}>
+            <div className={styles.projects_containers}>
+              <div className="cards">
+                <div className="article-container">
+                  <Image
+                    src={web1}
+                    alt="portfolio image 1"
+                    className="rounded-lg object-cover"
+                    width={"100%"} height={"100%"} />
+                </div>
+                <h2 className={styles.project_title}> Goals list app</h2>
+                <div className=" basis-1/2 flex-1 px-10">
+                <button
+                  className="btn_projects btn" 
+                  type="button" 
+                  onClick={(e) => {e.preventDefault();
+                    window.location.href='https://github.com/cs-mmeza/goals_app';
+                    }}>
+                    GitHub Repo
+                  </button>
+                  {/* <button
+                    className="btn_projects btn ml-10 md:ml-20 lg:ml-30"
+                    onClick="location.href='https://github.com/'"
+                  >
+                    Demo
+                  </button> */}
+                </div>
+              </div>
+              <div className="cards">
+                <div className="article-container">
+                  <Image
+                    src={web1}
+                    alt="portfolio image 1"
+                    className="rounded-lg object-cover"
+                    width={"100%"} height={"100%"} />
+                </div>
+                <h2 className={styles.project_title}> API Goals List </h2>
+                <div className=" basis-1/2 flex-1 px-10">
+                  <button
+                    className="btn_projects btn"
+                    onClick={(e) => {e.preventDefault();
+                      window.location.href='https://github.com/cs-mmeza/api_goals_app';
+                      }}>
+                    GitHub Repo
+                  </button>
+                  {/* <button
+                    className="btn_projects btn ml-10 md:ml-20 lg:ml-30"
+                    type="button"
+                    onClick={(e) => {e.preventDefault();
+                      window.location.href='https://github.com/cs-mmeza/api_goals_app';
+                      }}>
+                    Demo
+                  </button> */}
+                </div>
+              </div>
+              <div className="cards">
+                <div className="article-container">
+                  <Image
+                    src={web2}
+                    alt="portfolio image 1"
+                    className="rounded-lg object-cover"
+                    width={"100%"} height={"100%"} />
+                </div>
+                <h2 className={styles.project_title}>Business website</h2>
+                <div className=" basis-1/2 flex-1 px-10 items-center">
+                  <button
+                    className="btn_projects btn px-3"
+                    onClick={(e) => {e.preventDefault();
+                      window.location.href='https://github.com/cs-mmeza/norte-arq-demo';
+                      }}>
+                    GitHub Repo
+                  </button>
+                  {/* <button
+                    className="btn_projects btn ml-10 md:ml-20 lg:ml-30"
+                    onClick="location.href= https://github.com/"
+                  >
+                    Demo
+                  </button> */}
+                </div>
+              </div>
+            </div>
+          </div>
+        </section>
+      </main>
+    </div>
+  )
+}
